requirements.txt
flask
flask-cors
flask-mysqldb
python-dotenv
bcrypt
PyJWT
mysql-connector-python
